requirements.txt
langchain==0.2.5
langchain-openai==0.1.9
openai==1.35.3
openai-whisper==20231117
python-dotenv==1.0.1
setuptools-rust==1.9.0
tqdm==4.66.4
gradio==4.36.1